[SERVER] refactor: 아이템 인벤토리 관련 exception 추가

See merge request s10-final/S10P31C201!174

+++ server/realtime/src/main/java/com/explorer/realtime/gamedatahandling/inventory/event/DropItemInInventory.java
@@ -2,8 +2,11 @@ package com.explorer.realtime.gamedatahandling.inventory.event;
 
 import com.explorer.realtime.gamedatahandling.component.personal.inventoryInfo.dto.InventoryInfo;
 import com.explorer.realtime.gamedatahandling.component.personal.inventoryInfo.repository.InventoryRepository;
+import com.explorer.realtime.gamedatahandling.component.personal.playerInfo.repository.PlayerInfoRepository;
 import com.explorer.realtime.gamedatahandling.farming.repository.MapInfoRepository;
 import com.explorer.realtime.gamedatahandling.inventory.dto.ItemInfo;
+import com.explorer.realtime.gamedatahandling.inventory.exception.InventoryErrorCode;
+import com.explorer.realtime.gamedatahandling.inventory.exception.InventoryException;
 import com.explorer.realtime.global.common.dto.Message;
 import com.explorer.realtime.global.common.enums.CastingType;
 import com.explorer.realtime.global.component.broadcasting.Broadcasting;
@@ -25,6 +28,7 @@ public class DropItemInInventory {
 
     private final InventoryRepository inventoryRepository;
     private final MapInfoRepository mapInfoRepository;
+    private final PlayerInfoRepository playerInfoRepository;
     private final Unicasting unicasting;
     private final Broadcasting broadcasting;
 
@@ -35,34 +39,58 @@ public class DropItemInInventory {
         int mapId = json.getInt("mapId");
         Long userId = json.getLong("userId");
         int inventoryIdx = json.getInt("inventoryIdx");
-        int itemCnt = json.getInt("itemCnt");
-        String position = json.getString("position");
-        log.info("[process] channelId : {}, mapId : {} , userId : {}, inventoryIdx : {}, itemCnt : {}, position : {}", channelId, mapId, userId, inventoryIdx, itemCnt, position);
 
-        return inventoryRepository.findByInventoryIdx(channelId, userId, inventoryIdx)
+        String position = json.getString("position");
+        log.info("[process] channelId : {}, mapId : {} , userId : {}, inventoryIdx : {}, position : {}", channelId, mapId, userId, inventoryIdx, position);
+
+        return getInventoryCntByUserId(channelId, userId)
+                .flatMap(maxCnt -> {
+                    if (inventoryIdx < 0 || inventoryIdx >= maxCnt) {
+                        return Mono.error(new InventoryException(InventoryErrorCode.OUT_OF_RANGE_INDEX));
+                    }
+                    return Mono.empty();
+                })
+                .then(inventoryRepository.findByInventoryIdx(channelId, userId, inventoryIdx))
                 .flatMap(inventory -> {
+                    if (String.valueOf(inventory).isEmpty()) {
+                        return Mono.error(new InventoryException(InventoryErrorCode.EMPTY_INVENTORY));
+                    }
+
                     InventoryInfo inventoryInfo = InventoryInfo.ofString(inventoryIdx, String.valueOf(inventory));
                     String itemCategory = inventoryInfo.getItemCategory();
                     int itemId = inventoryInfo.getItemId();
                     log.info("[process] inventoryInfo : {}", inventoryInfo);
 
-                    int result = inventoryInfo.getItemCnt() - itemCnt;
+                    int result = inventoryInfo.getItemCnt() - 1;
                     if (result <= 0) {
                         return inventoryRepository.deleteByInventoryIdx(channelId, userId, inventoryIdx)
-                                .then(Mono.just(InventoryInfo.ofString(inventoryIdx, "")))
-                                .flatMap(emptyInventoryInfo -> {
-                                    return mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId, itemCnt)
-                                            .then(unicasting(channelId, userId, emptyInventoryInfo))
-                                            .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId, itemCnt)));
-                                });
+                                .then(mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId))
+                                .then(unicasting(channelId, userId, InventoryInfo.ofString(inventoryIdx, "")))
+                                .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId)));
                     } else {
                         inventoryInfo.setItemCnt(result);
                         return inventoryRepository.save(channelId, userId, inventoryInfo)
-                                .then(mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId, itemCnt))
+                                .then(mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId))
                                 .then(unicasting(channelId, userId, inventoryInfo))
-                                .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId, itemCnt)));
+                                .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId)));
                     }
-                });
+                })
+                .onErrorResume(InventoryException.class, error -> {
+                    log.info("[process] errorCode : {}, errorMessage : {}", error.getErrorCode(), error.getMessage());
+                    unicasting.unicasting(
+                            channelId,
+                            userId,
+                            MessageConverter.convert(Message.fail(eventName, CastingType.UNICASTING, String.valueOf(error.getErrorCode()), error.getMessage()))
+                    ).subscribe();
+                    return Mono.empty();
+                }).then();
+    }
+
+    private Mono<Integer> getInventoryCntByUserId(String channelId, Long userId) {
+        log.info("[getInventoryCntByUserId] channelId : {}, userId : {}", channelId, userId);
+
+        return playerInfoRepository.findInventoryCnt(channelId, userId)
+                .map(map -> Integer.parseInt(String.valueOf(map)));
     }
 
     public Mono<Void> unicasting(String channelId, Long userId, InventoryInfo inventoryInfo) {
